Removes react-powerplug from DigitBottomNavigation story
