+++ stories/bottom-navigation/digit-bottom-navigation.stories.js
@@ -1,27 +1,12 @@
 import React from "react";
 
-import {
-  withKnobs,
-  select,
-  text,
-  boolean,
-  number
-} from "@storybook/addon-knobs";
+import { withKnobs, boolean } from "@storybook/addon-knobs";
 import { storiesOf } from "@storybook/react";
-import { action } from "@storybook/addon-actions";
-import { linkTo } from "@storybook/addon-links";
 import { withReadme } from "storybook-readme";
 
-import { Value } from "react-powerplug";
-
-import IconOne from "@material-ui/icons/Android";
-
-import { DigitBottomNavigation } from "../../components";
 import DigitBottomNavigationReadme from "../../components/elements/digit-bottom-navigation/readme.md";
 import DigitProviders from "../../components/declaratives/digit-providers";
-
-const labels = ["Label One", "Label Two", "Label Three"];
-const icons = [<IconOne />, <IconOne />, <IconOne />];
+import StoryDigitBottomNavigation from "./StoryDigitBottomNavigation";
 
 const DigitBottomNavigationStory = storiesOf("Elements", module);
 
@@ -34,20 +19,7 @@ DigitBottomNavigationStory.add(
 
     return (
       <DigitProviders>
-        <Value
-          initial={0}
-          render={({ value, set }) => (
-            <DigitBottomNavigation
-              selected={value}
-              labels={labels}
-              icons={icons}
-              showLabels={showLabels}
-              onChange={selected => {
-                set(selected);
-              }}
-            />
-          )}
-        />
+        <StoryDigitBottomNavigation showLabels={showLabels} />
       </DigitProviders>
     );
   })
